% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{s1}
\alias{s1}
\title{Identifier of the first short-term outcome}
\usage{
s1(x)
}
\arguments{
\item{x}{the first short-term outcome}
}
\value{
returns the input if this function is used directly. See `details` for more explanation
}
\description{
Identifier of the first short-term outcome
}
\details{
This is a helperfunction used by \code{multipredict()} to identify the first
short-term outcome. If used directly it will only return the input. Therefore it should
not be called directly and only be used in the `formula` argument in \code{multipredict()}
to identify the variable that is the first short-term outcome. See `Examples` section of
\code{multipredict()} for more details.
}
