% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{s2}
\alias{s2}
\title{Identifier of the second short-term outcome}
\usage{
s2(x)
}
\arguments{
\item{x}{the second short-term outcome}
}
\value{
returns the input if this function is used directly. See `details` for more explanation
}
\description{
Identifier of the second short-term outcome
}
\details{
This is a helperfunction used by \code{multipredict()} to identify the second
short-term outcome. It should not be called directly. If used directly it will only return the input. Therefore it should
not be called directly and only be used in the `formula` argument in \code{multipredict()}
to identify the variable that is the second short-term outcome. See `example` section of
\code{multipredict()} for more details.
}
