% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{simulation}
\alias{simulation}
\title{Simulated data for landmulti package}
\format{
An object of class \code{data.frame} with 3000 rows and 6 columns.
}
\usage{
data(simulation)
}
\description{
This is a simulated data used as an example and test for \code{multipredict()} function. The data is
randomly generated by computer and has no practical meanning.
}
\keyword{datasets}
